Enhance transactions microservice: extend Category entity
Add new functionality to Category entity (CRUD, validation)
Update Transactions and Category entities relationships
Category now holds a collection of transactions

+++ transactions/src/main/java/com/mybudget/transactions/service/implementation/TransactionServiceImpl.java
@@ -9,19 +9,15 @@ import com.mybudget.transactions.exception.ResourceNotFoundException;
 import com.mybudget.transactions.repository.CategoryRepository;
 import com.mybudget.transactions.repository.TransactionRepository;
 import com.mybudget.transactions.service.TransactionService;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.Optional;
 
 import static com.mybudget.common.kafka.Topics.STREAMING_TRANSACTIONS_SAGA_STARTED_V1;
 
@@ -36,30 +32,27 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Transactional
     public void deleteAllByKeycloakSub(String keycloakSub) {
+        if (keycloakSub == null) {
+            throw new ResourceNotFoundException("KeycloakSub", "keycloakSub", "not found");
+        }
         transactionRepository.deleteByKeycloakSub(keycloakSub);
     }
 
     @Override
-    public Optional<Transaction> findTransaction(String keycloakSub, Long id) {
-        return Optional.ofNullable(transactionRepository.findTransactionByKeycloakSubAndId(keycloakSub, id)
-                .orElseThrow(() -> new ResourceNotFoundException("Transaction", "id", id.toString())));
-    }
-
-    @Override
-    @Transactional
-    public void deleteAllByUsername(String username) {
-        transactionRepository.deleteAllByUsername(username);
+    @Transactional(readOnly = true)
+    public Transaction findByKeycloakSubAndId(String keycloakSub, Long id) {
+        return transactionRepository.findTransactionByKeycloakSubAndId(keycloakSub, id)
+                .orElseThrow(() -> new ResourceNotFoundException(keycloakSub, "Transaction", id.toString()));
     }
 
     @Transactional
     @Override
     public Transaction createTransaction(String keycloakSub,
-                                         String username,
                                          BigDecimal amount,
                                          Long categoryId,
                                          String description) {
 
-        logger.info("Creating transaction for sub: {}, amount: {}",
+        logger.info("Creating transaction for sub={}, amount={}",
                 keycloakSub, amount);
 
         Category category = categoryRepository.findById(categoryId)
@@ -75,13 +68,15 @@ public class TransactionServiceImpl implements TransactionService {
         Transaction transaction = new Transaction();
         transaction.setStatus(TransactionStatus.PENDING);
         transaction.setKeycloakSub(keycloakSub);
-        transaction.setUsername(username);
+        transaction.setTransactionType(category.getTransactionType());
         transaction.setAmount(amount);
-        transaction.setCategory(category);
         transaction.setDescription(description);
 
-        transactionRepository.save(transaction);
-        logger.info("Creating transaction for sub:{}, ID: {}, type: {}, amount: {}",
+        category.addTransaction(transaction);
+        category.setCategoryBalance(category.getCategoryBalance().add(amount));
+        categoryRepository.save(category);
+
+        logger.info("Creating transaction for sub={}, ID={}, type={}, amount={}",
                 keycloakSub, transaction.getId(), transaction.getTransactionType(), transaction.getAmount());
 
         TransactionSagaStartEvent event = new TransactionSagaStartEvent(
@@ -91,33 +86,44 @@ public class TransactionServiceImpl implements TransactionService {
                 transaction.getTransactionType().name()
         );
         kafkaTemplate.send(STREAMING_TRANSACTIONS_SAGA_STARTED_V1, event);
-        logger.info("SagaStartEvent published to Orchestrator with sub:{}, ID: {}, type: {}, amount: {}",
+        logger.info("SagaStartEvent published to Orchestrator with sub={}, ID={}, type={}, amount={}",
                 keycloakSub, transaction.getId(), transaction.getTransactionType(), transaction.getAmount());
 
         return transaction;
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Transaction> findByTransactionType(String keycloakSub, TransactionType transactionType) {
         return transactionRepository.findByKeycloakSubAndTransactionType(keycloakSub, transactionType);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Transaction> findTransactions(String keycloakSub) {
         return transactionRepository.findTransactionsByKeycloakSub(keycloakSub);
     }
 
+    @Override
+    @Transactional(readOnly = true)
+    public List<Transaction> findByCategory(String keycloakSub, Long categoryId) {
+        return transactionRepository.findByCategoryIdAndKeycloakSub(categoryId, keycloakSub);
+    }
+
     @Transactional
     @Override
-    public boolean deleteTransaction(String keycloakSub, Long id) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String authHeader = "Bearer " + ((JwtAuthenticationToken) authentication).getToken().getTokenValue();
+    public void deleteTransaction(String keycloakSub, Long id) {
 
         Transaction transaction = transactionRepository
                 .findTransactionByKeycloakSubAndId(keycloakSub, id)
-                .orElseThrow(() -> new ResourceNotFoundException("Transaction", "id", id.toString()));
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("Transaction", "id", id.toString()));
 
-        transactionRepository.delete(transaction);
-        return true;
+        Category category = transaction.getCategory();
+
+        category.removeTransaction(transaction);
+        category.setCategoryBalance(category.getCategoryBalance().subtract(transaction.getAmount()));
+
+        categoryRepository.save(category);
     }
 }
